% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getDefaultIsThermallingFunction}
\alias{getDefaultIsThermallingFunction}
\title{getDefaultIsThermallingFunction}
\usage{
getDefaultIsThermallingFunction(totalAngle = 360)
}
\arguments{
\item{totalAngle}{the cumilative angle that is required to consider an trajectory thermaling}
}
\value{
a function is returned that based on a series of headings returns a logical value to indicate is a track is thermaling or not
}
\description{
getDefaultIsThermallingFunction
}
\examples{
fun<-getDefaultIsThermallingFunction(170)
fun(1:160)
fun(1:190)
}

