% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{getIsFocalPointFunction}
\alias{getIsFocalPointFunction}
\title{A function to generate isFocalPoint functions}
\usage{
getIsFocalPointFunction(isFocalPoint)
}
\arguments{
\item{isFocalPoint}{a function, a boolean array from which such a function can be built, or a list of indices}
}
\value{
a function which decides if wind estimation is performed for a point in the input data
}
\description{
A function to generate isFocalPoint functions
}
