% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{getIsSamplingRegularFunction}
\alias{getIsSamplingRegularFunction}
\title{A function to generate functions used to check if a segment is regular}
\usage{
getIsSamplingRegularFunction(isSamplingRegular)
}
\arguments{
\item{isSamplingRegular}{a function which decides if a sequence of timestamps is regular or the interval which is considered regular}
}
\value{
a function which decides if a sequence of timestamps is regular
}
\description{
A function to generate functions used to check if a segment is regular
}
\examples{
fun<-getIsSamplingRegularFunction(10)
fun(Sys.time()+1:5)
fun(Sys.time()+c(0,10,20,30))
fun(Sys.time()+c(0,10,20,31))
}

