% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getSamplingIsRegularFunction}
\alias{getSamplingIsRegularFunction}
\title{getSamplingIsRegularFunction}
\usage{
getSamplingIsRegularFunction(samplingIntervalSeconds)
}
\arguments{
\item{samplingIntervalSeconds}{the interval that is considered regular}
}
\value{
a function is returned that based on a series of timestamps decides if the segment is regular
}
\description{
getSamplingIsRegularFunction
}
\examples{
fun<-getSamplingIsRegularFunction(10)
fun(Sys.time()+1:5)
fun(Sys.time()+c(0,10,20,30))
fun(Sys.time()+c(0,10,20,31))
}

