% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{getWindowSizeLR}
\alias{getWindowSizeLR}
\title{Generate arguments for window size around focal point}
\usage{
getWindowSizeLR(windowSize)
}
\arguments{
\item{windowSize}{a pair of positive integers determining the window size left and right of a focal point or an odd number determining the size of a symmetrical window}
}
\value{
windowSize a pair of positive integers determining the window size left and right of a focal point
}
\description{
A function to translate an window size argument to a standardized argument.
}

