% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{windEstimLogLik}
\alias{windEstimLogLik}
\title{Estimate the log likelihood}
\usage{
windEstimLogLik(sigma, phi)
}
\arguments{
\item{sigma}{the residual variance in airspeed}

\item{phi}{the autocorrelation used in the calculations}
}
\value{
the log likelihood
}
\description{
Estimate the log likelihood
}
\examples{
windEstimLogLik(c(1.3,.6,1.5,1.8),.3)
windEstimLogLik(c(1.3,.6,1.5,1.8),.5)
}

