% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storks.R
\docType{data}
\name{storks}
\alias{storks}
\title{Example stork data.}
\format{A MoveStack consisting of 22333 locations}
\source{
\url{http://www.movebank.org/}
}
\usage{
storks
}
\description{
A dataset containing location data of 6 juvenile storks (Ciconia ciconia) on the 18th of august when migration just started. On several occasion the birds use thermals.
}
\examples{
data("storks")
}
\keyword{datasets}
